@@ -14,32 +14,37 @@
   "author": "",
   "license": "GPLv3",
   "dependencies": {
-    "@emotion/react": "^11.4.0",
-    "@emotion/server": "^11.4.0",
-    "@emotion/styled": "^11.3.0",
-    "@material-ui/core": "^5.0.0-alpha.34",
-    "@material-ui/icons": "^4.11.2",
-    "axios": "^0.21.4",
+    "@emotion/react": "^11.10.5",
+    "@emotion/server": "^11.10.0",
+    "@emotion/styled": "^11.10.5",
+    "@fontsource/roboto": "^4.5.8",
+    "@mui/icons-material": "^5.10.16",
+    "@mui/material": "^5.10.16",
+    "@mui/styled-engine-sc": "^5.10.16",
+    "axios": "^1.2.0",
     "geobuf": "^3.0.2",
     "js-yaml": "^4.1.0",
-    "leaflet": "^1.7.1",
-    "leaflet.fullscreen": "^2.0.0",
-    "next": "^11.1.3",
+    "leaflet": "^1.9.3",
+    "leaflet.fullscreen": "^2.4.0",
+    "next": "^13.0.6",
     "pbf": "^3.2.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-infinite-scroller": "^1.2.6",
-    "react-leaflet": "^3.2.0",
+    "react-leaflet": "^4.2.0",
     "react-swipeable-views": "^0.14.0",
     "react-swipeable-views-utils": "^0.14.0",
-    "screenfull": "^5.1.0"
+    "react-window": "^1.8.8",
+    "screenfull": "^6.0.2",
+    "styled-components": "^5.3.6"
   },
   "devDependencies": {
-    "@types/leaflet": "^1.7.3",
-    "@types/react": "^17.0.11",
-    "eslint": "^7.29.0",
-    "eslint-config-next": "^11.0.0",
-    "prettier": "2.3.1",
-    "typescript": "^4.3.4"
+    "@types/leaflet": "^1.9.0",
+    "@types/node": "18.11.10",
+    "@types/react": "^18.0.26",
+    "eslint": "^8.29.0",
+    "eslint-config-next": "^13.0.6",
+    "prettier": "2.8.0",
+    "typescript": "^4.9.3"
   }
 }
