@@ -1,5 +1,5 @@
-import { Typography } from '@material-ui/core'
-import { Stack } from '@material-ui/core'
+import { Typography } from '@mui/material'
+import { Stack } from '@mui/material'
 import { FC } from 'react'
 
 export interface FormItemProps {
@@ -11,7 +11,9 @@ const FormItem: FC<FormItemProps> = ({ label, value }) => {
   return (
     <Stack>
       <Typography>{label}</Typography>
-      <Typography variant="caption" style={{ whiteSpace: 'pre-wrap' }} >{value}</Typography>
+      <Typography variant="caption" style={{ whiteSpace: 'pre-wrap' }}>
+        {value}
+      </Typography>
     </Stack>
   )
 }
