@@ -1,24 +1,16 @@
-import { Container, makeStyles } from '@material-ui/core'
+import { Container, makeStyles } from '@mui/material'
 import { FC } from 'react'
 import Footer from '../Footer'
 import NavBar from '../NavBar'
 
-const useStyles = makeStyles((theme) => ({
-  root: {
-    marginTop: 56,
-  },
-  [theme.breakpoints.up('sm')]: {
-    root: {
-      marginTop: 62,
-    },
-  },
-}))
 
-const LayoutDefault: FC = ({ children }) => {
-  const classes = useStyles()
+interface LayoutDefaultProps {
+  children: JSX.Element | JSX.Element[]
+}
 
+const LayoutDefault: FC<LayoutDefaultProps> = ({ children }) => {
   return (
-    <Container className={classes.root} maxWidth={false} disableGutters={true}>
+    <Container maxWidth={false} disableGutters={true}>
       <NavBar />
       <main>{children}</main>
       <Footer />
